Start parsing from 0.
The method parse(argc, argv) should start parsing from
argv[0]. Adjustment should be made when it's called. So if you are
calling from the main(argc, argv), you probably want to call
```.parse(argc-1, argv+1)```. But if you want to call on the rest of the
args after --, make no adjustment.